--- src/src/components/Navbar.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Navbar = () => {
-  return (
-    <nav className='navbar navbar-light bg-dark'>
-      <h3>ROI Calculator</h3>
-    </nav>
-  );
-};
-
-export default Navbar;
